pyproject.toml
[project]
name = "flafludeagentes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.8.0",
    "google-generativeai>=0.8.5",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
    "typing-extensions>=4.14.1",
    "flask>=3.0.0",
    "httpx>=0.28.1",
]
